Corrected PDF calculation in README.rdoc
The correct PDF of a normal distribution with mean = 10 and sd = 2 at x = 11 is 0.1760. The reason why it was wrong before seems to be some nasty integer calculation. I ran into this problem using the rubystats normal distribution with my own monte carlo code. I suggest we adjust the code for all continuous distributions to make an explicit conversion of the input parameters to floating point numbers at every initialize method.